refactor: consolidate shared source and adapter contracts
