+++ .changeset/shared-framework-contracts.md
@@ -1,0 +1,11 @@
+---
+"@pracht/capabilities": patch
+"@pracht/core": patch
+"@pracht/cli": patch
+"@pracht/vite-plugin": patch
+"@pracht/adapter-node": patch
+"@pracht/adapter-cloudflare": patch
+"@pracht/adapter-netlify": patch
+---
+
+Keep page metadata, environment checks, route inspection, and prerendered headers consistent across framework tools and adapters.
