+++ packages/cli/src/verification-env.ts
@@ -4,16 +4,12 @@ import { join, resolve } from "node:path";
 import { displayPath, listFilesRecursively, type ProjectConfig } from "./project.js";
 import { createCheck, type Check } from "./verification-helpers.js";
 
-// Keep in sync with packages/vite-plugin/src/env-safety.ts — the CLI cannot
-// depend on @pracht/vite-plugin, so the (small) scan logic is mirrored here.
-const VITE_BUILTIN_ENV_VARS = new Set(["MODE", "DEV", "PROD", "SSR", "BASE_URL", "NODE_ENV"]);
-const PUBLIC_ENV_PREFIX = "PRACHT_PUBLIC_";
-const ENV_REFERENCE_RE =
-  /\b(process\.env|import\.meta\.env)(?:\??\.([A-Za-z_$][A-Za-z0-9_$]*)|(?:\?\.)?\[\s*(["'])([A-Za-z_$][A-Za-z0-9_$]*)\3\s*\])/g;
-// `import.meta.env` reads that are not a single-key access; Vite replaces those
-// with an object literal holding every exposed variable.
-const WHOLE_ENV_READ_RE = /\bimport\.meta\.env\b(?!\s*\??\.\s*[A-Za-z_$])/g;
-const WHOLE_ENV_READ = "*";
+import {
+  getCodePositionMask,
+  WHOLE_ENV_READ,
+  scanCodeForEnvLeaks as scanSourceForEnvLeaks,
+} from "@pracht/capabilities/static";
+export { scanCodeForEnvLeaks as scanSourceForEnvLeaks } from "@pracht/capabilities/static";
 
 export interface EnvLeakFinding {
   accessor: string;
@@ -28,239 +24,6 @@ interface BuildEnvSafetyReport {
     name?: unknown;
     sources?: unknown;
   }>;
-}
-
-export function scanSourceForEnvLeaks(
-  code: string,
-  allow: ReadonlySet<string>,
-): { accessor: string; name: string }[] {
-  const codePositions = getCodePositionMask(code);
-  const matches: Array<{ accessor: string; index: number; name: string }> = [];
-
-  for (const match of code.matchAll(ENV_REFERENCE_RE)) {
-    const index = match.index ?? -1;
-    if (!codePositions[index]) continue;
-    const accessor = match[1];
-    const name = match[2] ?? match[4];
-    if (!name) continue;
-    if (name.startsWith(PUBLIC_ENV_PREFIX)) continue;
-    if (VITE_BUILTIN_ENV_VARS.has(name)) continue;
-    if (allow.has(name)) continue;
-    matches.push({ accessor, index, name });
-  }
-
-  if (!allow.has(WHOLE_ENV_READ)) {
-    for (const match of code.matchAll(WHOLE_ENV_READ_RE)) {
-      const index = match.index ?? -1;
-      if (!codePositions[index]) continue;
-      matches.push({ accessor: "import.meta.env", index, name: WHOLE_ENV_READ });
-    }
-  }
-
-  const findings: { accessor: string; name: string }[] = [];
-  const seen = new Set<string>();
-  for (const { accessor, name } of matches.sort((a, b) => a.index - b.index)) {
-    const key = `${accessor}.${name}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    findings.push({ accessor, name });
-  }
-
-  return findings;
-}
-
-function getCodePositionMask(code: string): Uint8Array {
-  const mask = new Uint8Array(code.length);
-  const templateExpressionDepths: number[] = [];
-  let mode: "block-comment" | "code" | "double" | "line-comment" | "regex" | "single" | "template" =
-    "code";
-  let regexCharClass = false;
-  let i = 0;
-
-  while (i < code.length) {
-    const char = code[i];
-    const next = code[i + 1];
-
-    if (mode === "line-comment") {
-      if (char === "\n" || char === "\r") {
-        mode = "code";
-        mask[i] = 1;
-      }
-      i++;
-      continue;
-    }
-
-    if (mode === "block-comment") {
-      if (char === "*" && next === "/") {
-        mode = "code";
-        i += 2;
-      } else {
-        i++;
-      }
-      continue;
-    }
-
-    if (mode === "single" || mode === "double") {
-      const quote = mode === "single" ? "'" : '"';
-      if (char === "\\") {
-        i += 2;
-        continue;
-      }
-      if (char === quote || char === "\n" || char === "\r") {
-        mode = "code";
-      }
-      i++;
-      continue;
-    }
-
-    if (mode === "regex") {
-      if (char === "\\") {
-        i += 2;
-        continue;
-      }
-      if (char === "[") {
-        regexCharClass = true;
-        i++;
-        continue;
-      }
-      if (char === "]") {
-        regexCharClass = false;
-        i++;
-        continue;
-      }
-      if (char === "/" && !regexCharClass) {
-        regexCharClass = false;
-        i++;
-        while (i < code.length && isIdentifierChar(code[i])) i++;
-        mode = "code";
-        continue;
-      }
-      if (char === "\n" || char === "\r") {
-        regexCharClass = false;
-        mode = "code";
-      }
-      i++;
-      continue;
-    }
-
-    if (mode === "template") {
-      if (char === "\\") {
-        i += 2;
-        continue;
-      }
-      if (char === "`") {
-        mode = "code";
-        i++;
-        continue;
-      }
-      if (char === "$" && next === "{") {
-        mask[i] = 1;
-        mask[i + 1] = 1;
-        templateExpressionDepths.push(1);
-        mode = "code";
-        i += 2;
-        continue;
-      }
-      i++;
-      continue;
-    }
-
-    mask[i] = 1;
-
-    if (char === "/" && next === "/") {
-      mask[i + 1] = 1;
-      mode = "line-comment";
-      i += 2;
-      continue;
-    }
-
-    if (char === "/" && next === "*") {
-      mask[i + 1] = 1;
-      mode = "block-comment";
-      i += 2;
-      continue;
-    }
-
-    if (char === "/" && isRegexLiteralStart(code, i)) {
-      mode = "regex";
-      regexCharClass = false;
-      i++;
-      continue;
-    }
-
-    if (char === "'") {
-      mode = "single";
-      i++;
-      continue;
-    }
-
-    if (char === '"') {
-      mode = "double";
-      i++;
-      continue;
-    }
-
-    if (char === "`") {
-      mode = "template";
-      i++;
-      continue;
-    }
-
-    if (templateExpressionDepths.length > 0) {
-      const top = templateExpressionDepths.length - 1;
-      if (char === "{") {
-        templateExpressionDepths[top]++;
-      } else if (char === "}") {
-        templateExpressionDepths[top]--;
-        if (templateExpressionDepths[top] === 0) {
-          templateExpressionDepths.pop();
-          mode = "template";
-        }
-      }
-    }
-
-    i++;
-  }
-
-  return mask;
-}
-
-function isRegexLiteralStart(code: string, slashIndex: number): boolean {
-  let i = slashIndex - 1;
-  while (i >= 0 && /\s/.test(code[i])) i--;
-  if (i < 0) return true;
-
-  const previous = code[i];
-  if (previous === ">" && code[i - 1] === "=") return true;
-  if ("([{=,:;!?&|^~<>*%+-".includes(previous)) return true;
-
-  if (isIdentifierChar(previous)) {
-    let start = i;
-    while (start >= 0 && isIdentifierChar(code[start])) start--;
-    const word = code.slice(start + 1, i + 1);
-    return new Set([
-      "await",
-      "case",
-      "delete",
-      "do",
-      "else",
-      "in",
-      "instanceof",
-      "new",
-      "of",
-      "return",
-      "throw",
-      "typeof",
-      "void",
-      "yield",
-    ]).has(word);
-  }
-
-  return false;
-}
-
-function isIdentifierChar(char: string | undefined): boolean {
-  return !!char && /[A-Za-z0-9_$]/.test(char);
 }
 
 /**
