+++ packages/framework/src/runtime-headers.ts
@@ -259,3 +259,34 @@ export function appendVaryHeader(headers: Headers, value: string): void {
 
   headers.set("vary", `${current}, ${value}`);
 }
+
+/** Headers recorded while prerendering, indexed by route pathname. */
+export type HeadersManifest = Record<string, Record<string, string>>;
+
+export function applyHeadersManifest(
+  headers: Headers,
+  headersManifest: HeadersManifest,
+  pathname: string,
+): void {
+  const routeHeaders = getManifestHeaders(headersManifest, pathname);
+  if (!routeHeaders) return;
+
+  for (const [key, value] of Object.entries(routeHeaders)) {
+    headers.set(key, value);
+  }
+}
+
+export function getManifestHeaders(
+  headersManifest: HeadersManifest,
+  pathname: string,
+): Record<string, string> | undefined {
+  const withoutIndex = pathname.replace(/\/index\.html$/, "") || "/";
+  const withoutSlash = pathname.replace(/\/$/, "") || "/";
+
+  return (
+    headersManifest[pathname] ??
+    headersManifest[withoutSlash] ??
+    headersManifest[withoutIndex] ??
+    undefined
+  );
+}
