+++ packages/vite-plugin/src/env-safety.ts
@@ -1,21 +1,18 @@
 import type { Plugin } from "vite";
-
-/**
- * Env vars Vite defines on `import.meta.env` in every bundle, plus NODE_ENV
- * which Vite's define pass statically replaces at build time (so it can never
- * leak and is referenced by countless dependencies).
- */
-export const VITE_BUILTIN_ENV_VARS = new Set([
-  "MODE",
-  "DEV",
-  "PROD",
-  "SSR",
-  "BASE_URL",
-  "NODE_ENV",
-]);
-
-/** Prefix that marks an env var as intentionally public. */
-export const PUBLIC_ENV_PREFIX = "PRACHT_PUBLIC_";
+import {
+  PUBLIC_ENV_PREFIX,
+  VITE_BUILTIN_ENV_VARS,
+  WHOLE_ENV_READ,
+  scanCodeForEnvLeaks,
+} from "@pracht/capabilities/static";
+import type { EnvLeakReference } from "@pracht/capabilities/static";
+export {
+  PUBLIC_ENV_PREFIX,
+  VITE_BUILTIN_ENV_VARS,
+  WHOLE_ENV_READ,
+  scanCodeForEnvLeaks,
+} from "@pracht/capabilities/static";
+export type { EnvLeakReference } from "@pracht/capabilities/static";
 
 /** Server-only core entry that must never resolve into client bundles. */
 export const SERVER_ENV_MODULE_ID = "@pracht/core/env/server";
@@ -28,270 +25,6 @@ export interface EnvSafetyReport {
 export interface EnvSafetyOptions {
   /** Env var names allowed to appear in client bundles despite not being public. */
   allow?: string[];
-}
-
-export interface EnvLeakReference {
-  accessor: "process.env" | "import.meta.env";
-  name: string;
-}
-
-/**
- * Sentinel `name` for an `import.meta.env` read that is not a static single-key
- * access. Vite only narrows `import.meta.env.KEY` / `import.meta.env?.KEY`
- * member expressions to their value; every other read — a bare reference,
- * destructuring, a spread, or bracket access — is replaced by an object literal
- * holding every exposed variable, so the `VITE_` values Pracht treats as
- * non-public end up verbatim in the client bundle. Those reads leave no
- * accessor text behind, so the name-based scan below cannot see them.
- */
-export const WHOLE_ENV_READ = "*";
-
-// Matches `process.env.X`, `import.meta.env.X`, their optional-chained forms
-// (`import.meta.env?.X`), and the equivalent bracket-string forms
-// (`process.env["X"]`, `import.meta.env['X']`).
-const ENV_REFERENCE_RE =
-  /\b(process\.env|import\.meta\.env)(?:\??\.([A-Za-z_$][A-Za-z0-9_$]*)|(?:\?\.)?\[\s*(["'])([A-Za-z_$][A-Za-z0-9_$]*)\3\s*\])/g;
-
-// Matches `import.meta.env` that is *not* followed by a single-key access, i.e.
-// exactly the reads Vite materializes into the full env object.
-const WHOLE_ENV_READ_RE = /\bimport\.meta\.env\b(?!\s*\??\.\s*[A-Za-z_$])/g;
-
-/**
- * Scans JavaScript source for references to environment variables that are
- * neither public-prefixed, Vite built-ins, nor explicitly allowed, plus reads
- * that pull in the whole `import.meta.env` object.
- */
-export function scanCodeForEnvLeaks(
-  code: string,
-  allow: ReadonlySet<string> = new Set(),
-): EnvLeakReference[] {
-  const codePositions = getCodePositionMask(code);
-  const matches: Array<{ index: number; reference: EnvLeakReference }> = [];
-
-  for (const match of code.matchAll(ENV_REFERENCE_RE)) {
-    const index = match.index ?? -1;
-    if (!codePositions[index]) continue;
-    const accessor = match[1] as EnvLeakReference["accessor"];
-    const name = match[2] ?? match[4];
-    if (!name) continue;
-    if (name.startsWith(PUBLIC_ENV_PREFIX)) continue;
-    if (VITE_BUILTIN_ENV_VARS.has(name)) continue;
-    if (allow.has(name)) continue;
-    matches.push({ index, reference: { accessor, name } });
-  }
-
-  if (!allow.has(WHOLE_ENV_READ)) {
-    for (const match of code.matchAll(WHOLE_ENV_READ_RE)) {
-      const index = match.index ?? -1;
-      if (!codePositions[index]) continue;
-      matches.push({ index, reference: { accessor: "import.meta.env", name: WHOLE_ENV_READ } });
-    }
-  }
-
-  const findings: EnvLeakReference[] = [];
-  const seen = new Set<string>();
-  for (const { reference } of matches.sort((a, b) => a.index - b.index)) {
-    const key = `${reference.accessor}.${reference.name}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    findings.push(reference);
-  }
-
-  return findings;
-}
-
-function getCodePositionMask(code: string): Uint8Array {
-  const mask = new Uint8Array(code.length);
-  const templateExpressionDepths: number[] = [];
-  let mode: "block-comment" | "code" | "double" | "line-comment" | "regex" | "single" | "template" =
-    "code";
-  let regexCharClass = false;
-  let i = 0;
-
-  while (i < code.length) {
-    const char = code[i];
-    const next = code[i + 1];
-
-    if (mode === "line-comment") {
-      if (char === "\n" || char === "\r") {
-        mode = "code";
-        mask[i] = 1;
-      }
-      i++;
-      continue;
-    }
-
-    if (mode === "block-comment") {
-      if (char === "*" && next === "/") {
-        mode = "code";
-        i += 2;
-      } else {
-        i++;
-      }
-      continue;
-    }
-
-    if (mode === "single" || mode === "double") {
-      const quote = mode === "single" ? "'" : '"';
-      if (char === "\\") {
-        i += 2;
-        continue;
-      }
-      if (char === quote || char === "\n" || char === "\r") {
-        mode = "code";
-      }
-      i++;
-      continue;
-    }
-
-    if (mode === "regex") {
-      if (char === "\\") {
-        i += 2;
-        continue;
-      }
-      if (char === "[") {
-        regexCharClass = true;
-        i++;
-        continue;
-      }
-      if (char === "]") {
-        regexCharClass = false;
-        i++;
-        continue;
-      }
-      if (char === "/" && !regexCharClass) {
-        regexCharClass = false;
-        i++;
-        while (i < code.length && isIdentifierChar(code[i])) i++;
-        mode = "code";
-        continue;
-      }
-      if (char === "\n" || char === "\r") {
-        regexCharClass = false;
-        mode = "code";
-      }
-      i++;
-      continue;
-    }
-
-    if (mode === "template") {
-      if (char === "\\") {
-        i += 2;
-        continue;
-      }
-      if (char === "`") {
-        mode = "code";
-        i++;
-        continue;
-      }
-      if (char === "$" && next === "{") {
-        mask[i] = 1;
-        mask[i + 1] = 1;
-        templateExpressionDepths.push(1);
-        mode = "code";
-        i += 2;
-        continue;
-      }
-      i++;
-      continue;
-    }
-
-    mask[i] = 1;
-
-    if (char === "/" && next === "/") {
-      mask[i + 1] = 1;
-      mode = "line-comment";
-      i += 2;
-      continue;
-    }
-
-    if (char === "/" && next === "*") {
-      mask[i + 1] = 1;
-      mode = "block-comment";
-      i += 2;
-      continue;
-    }
-
-    if (char === "/" && isRegexLiteralStart(code, i)) {
-      mode = "regex";
-      regexCharClass = false;
-      i++;
-      continue;
-    }
-
-    if (char === "'") {
-      mode = "single";
-      i++;
-      continue;
-    }
-
-    if (char === '"') {
-      mode = "double";
-      i++;
-      continue;
-    }
-
-    if (char === "`") {
-      mode = "template";
-      i++;
-      continue;
-    }
-
-    if (templateExpressionDepths.length > 0) {
-      const top = templateExpressionDepths.length - 1;
-      if (char === "{") {
-        templateExpressionDepths[top]++;
-      } else if (char === "}") {
-        templateExpressionDepths[top]--;
-        if (templateExpressionDepths[top] === 0) {
-          templateExpressionDepths.pop();
-          mode = "template";
-        }
-      }
-    }
-
-    i++;
-  }
-
-  return mask;
-}
-
-function isRegexLiteralStart(code: string, slashIndex: number): boolean {
-  let i = slashIndex - 1;
-  while (i >= 0 && /\s/.test(code[i])) i--;
-  if (i < 0) return true;
-
-  const previous = code[i];
-  if (previous === ">" && code[i - 1] === "=") return true;
-  if ("([{=,:;!?&|^~<>*%+-".includes(previous)) return true;
-
-  if (isIdentifierChar(previous)) {
-    let start = i;
-    while (start >= 0 && isIdentifierChar(code[start])) start--;
-    const word = code.slice(start + 1, i + 1);
-    return new Set([
-      "await",
-      "case",
-      "delete",
-      "do",
-      "else",
-      "in",
-      "instanceof",
-      "new",
-      "of",
-      "return",
-      "throw",
-      "typeof",
-      "void",
-      "yield",
-    ]).has(word);
-  }
-
-  return false;
-}
-
-function isIdentifierChar(char: string | undefined): boolean {
-  return !!char && /[A-Za-z0-9_$]/.test(char);
 }
 
 interface EnvLeakProblem extends EnvLeakReference {
