+++ packages/vite-plugin/src/pages-router.ts
@@ -1,12 +1,23 @@
-import { readdirSync, readFileSync, statSync, writeFileSync } from "node:fs";
-import { basename, dirname, extname, join, relative, resolve } from "node:path";
 import {
+  readPageStringExport,
+  readPageRevalidation,
+  PAGES_APP_CONFIG_EXPORTS,
+  pagesShellName,
+  findOwningPagesShell,
+  maskMarkdownFences,
   hasNamedMiddlewareExport,
   hasNamedValueExport,
   hasValueStarExport,
-  maskCommentsAndStrings,
   resolvePagesCapabilityName,
 } from "@pracht/capabilities/static";
+export {
+  PAGES_APP_CONFIG_EXPORTS,
+  pagesShellName,
+  findOwningPagesShell,
+} from "@pracht/capabilities/static";
+import { readdirSync, readFileSync, statSync, writeFileSync } from "node:fs";
+import { basename, dirname, extname, join, relative, resolve } from "node:path";
+
 import { parseAst } from "vite";
 import { getRolldownLang } from "./client-module-query.ts";
 import { detectHeadExport, detectHeadersExport, detectLoaderExport } from "./route-loader-hints.ts";
@@ -150,9 +161,6 @@ export interface PagesAppShell {
 }
 
 /** The registered shell name for an `_app` in `directory` (posix, `""` at the root). */
-export function pagesShellName(directory: string): string {
-  return directory === "" ? "pages" : `pages:${directory}`;
-}
 
 /**
  * Every `_app` shell in a pages directory, deepest first.
@@ -211,17 +219,6 @@ export function findPagesAppShellFiles(
 }
 
 /** The `_app` that owns a page: the nearest ancestor directory with one. */
-export function findOwningPagesShell(
-  shells: readonly PagesAppShell[],
-  pageRelativePath: string,
-): PagesAppShell | undefined {
-  const segments = pageRelativePath.replace(/\\/g, "/").split("/").slice(0, -1);
-  return shells.find(
-    (shell) =>
-      shell.directory === "" ||
-      segments.slice(0, shell.directory.split("/").length).join("/") === shell.directory,
-  );
-}
 
 /** Whether a middleware module explicitly exports, or may re-export, `middleware`. */
 function exportsMiddleware(source: string, file: string): boolean {
@@ -248,7 +245,6 @@ export const PAGES_APP_CONFIG_BASENAME = "_app.config";
  * and its Web Bot Auth keys — into the browser bundle. `pages/404.tsx` is
  * already the file convention for it.
  */
-export const PAGES_APP_CONFIG_EXPORTS = ["agents", "constraints"] as const;
 
 export interface PagesAppConfig {
   absolutePath: string;
@@ -533,119 +529,29 @@ function extractQuotedPageExport(
   name: "RENDER_MODE" | "HYDRATION",
   relativePath: string,
 ): string | undefined {
-  const masked = maskCommentsAndStrings(source);
-  const declarations = [...masked.matchAll(new RegExp(`export\\s+const\\s+${name}\\s*=`, "g"))];
-  if (declarations.length === 0) return undefined;
-  if (declarations.length > 1) {
+  const { count, value } = readPageStringExport(source, name);
+  if (count > 1)
     throw new Error(
       `[pracht] Pages route ${JSON.stringify(relativePath)} exports ${name} more than once.`,
     );
-  }
-
-  const declaration = declarations[0];
-  const valueStart = (declaration.index ?? 0) + declaration[0].length;
-  return source
-    .slice(valueStart)
-    .trimStart()
-    .match(/^["'](\w+)["']/)?.[1];
+  return value;
 }
-
-const REVALIDATE_RE = /export\s+const\s+REVALIDATE\s*=\s*([^;\n]+)/;
 
 function extractRevalidateSeconds(
   source: string,
   relativePath: string,
 ): { present: boolean; seconds?: number } {
-  const matches = [...maskCommentsAndStrings(source).matchAll(new RegExp(REVALIDATE_RE, "g"))];
-  if (matches.length === 0) return { present: false };
-  if (matches.length > 1) {
-    throw new Error(
-      `[pracht] Pages route ${JSON.stringify(relativePath)} exports REVALIDATE more than once.`,
-    );
-  }
-  const match = matches[0];
-
-  const expression = match[1].trim().replace(/\s+as\s+const$/, "");
-  if (!/^\d(?:_?\d)*$/.test(expression)) {
-    throw new Error(
-      `[pracht] Pages route ${JSON.stringify(relativePath)} must export REVALIDATE as a ` +
-        "positive integer literal number of seconds (for example, `export const REVALIDATE = 60`).",
-    );
-  }
-
-  const seconds = Number(expression.replaceAll("_", ""));
-  if (!Number.isSafeInteger(seconds) || seconds <= 0) {
-    throw new Error(
-      `[pracht] Pages route ${JSON.stringify(relativePath)} must export REVALIDATE as a ` +
-        "positive integer literal number of seconds within JavaScript's safe integer range.",
-    );
-  }
-
-  return { present: true, seconds };
-}
-
-/** Mask Markdown fenced examples while preserving source offsets and top-level MDX exports. */
-function maskMarkdownFences(source: string, relativePath: string): string {
-  if (!/\.mdx?$/.test(relativePath)) return source;
-
-  const chars = source.split("");
-  let activeFence: { character: "`" | "~"; continuationIndent: number; length: number } | null =
-    null;
-  for (const line of source.matchAll(/.*(?:\r?\n|$)/g)) {
-    if (line[0] === "") continue;
-    const lineStart = line.index ?? 0;
-    const content = line[0].replace(/\r?\n$/, "");
-    const stripped = stripMarkdownContainerPrefix(content);
-    const fenceContent: string =
-      activeFence && stripped.content.startsWith(" ".repeat(activeFence.continuationIndent))
-        ? stripped.content.slice(activeFence.continuationIndent)
-        : stripped.content;
-    const opening: RegExpExecArray | null = activeFence
-      ? null
-      : /^ {0,3}(`{3,}|~{3,})/.exec(fenceContent);
-    const closing = activeFence
-      ? new RegExp(`^ {0,3}\\${activeFence.character}{${activeFence.length},}[ \\t]*$`).test(
-          fenceContent,
-        )
-      : false;
-
-    if (activeFence || opening) {
-      for (let offset = 0; offset < line[0].length; offset += 1) {
-        const index = lineStart + offset;
-        if (chars[index] !== "\n" && chars[index] !== "\r") chars[index] = " ";
-      }
-    }
-
-    if (closing) {
-      activeFence = null;
-    } else if (opening) {
-      activeFence = {
-        character: opening[1][0] as "`" | "~",
-        continuationIndent: stripped.continuationIndent,
-        length: opening[1].length,
-      };
-    }
-  }
-  return chars.join("");
-}
-
-function stripMarkdownContainerPrefix(line: string): {
-  content: string;
-  continuationIndent: number;
-} {
-  let content = line;
-  let continuationIndent = 0;
-  while (true) {
-    const quote = /^ {0,3}> ?/.exec(content);
-    if (quote) {
-      content = content.slice(quote[0].length);
-      continue;
-    }
-    const list = /^ {0,3}(?:[-+*]|\d{1,9}[.)])[ \t]+/.exec(content);
-    if (!list) return { content, continuationIndent };
-    continuationIndent += list[0].length;
-    content = content.slice(list[0].length);
-  }
+  const result = readPageRevalidation(source);
+  if (result.kind === "missing") return { present: false };
+  if (result.kind === "time") return { present: true, seconds: result.seconds };
+  const prefix = `[pracht] Pages route ${JSON.stringify(relativePath)}`;
+  if (result.reason === "duplicate")
+    throw new Error(`${prefix} exports REVALIDATE more than once.`);
+  const detail =
+    result.reason === "range"
+      ? "positive integer literal number of seconds within JavaScript's safe integer range."
+      : "positive integer literal number of seconds (for example, `export const REVALIDATE = 60`).";
+  throw new Error(`${prefix} must export REVALIDATE as a ${detail}`);
 }
 
 export function generatePagesManifestSource(
